pyproject.toml
[tool.poetry]
name = "etl-backup-case"
version = "0.1.0"
description = ""
authors = ["Luciano Vasconcelos <lvgalvaofilho@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.5"
boto3 = "^1.35.4"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
